Move EPSFFitter to epsf module
